Implemented setter for mass and radius, optimised spawning.
There are now setters for mass and radius of every particle in a group.
The API allows to use them intuitively, i.e. just tell the mass of one
liter of particles. Also renamed spawnRect() to spawnCuboid() and
implemented the number of actual spawned particles as return value.
The example project uses all new functions for demonstration.